* Add game template

* Add quit button and fix overflow in question body

* Create Quiz, Question, Choice, and Player models

* Map urls and views (with test index)

* Edit Quiz's attribute name

* Add quizs, questions, choices in the admin site

* Create urls to game page (test version)

* Move quit button to the top right corner and display quiz topic from views

* Add status attributes to player and add method to quiz

* Add empty test file

* Display player, quiz and its question and choices on game.html

* Car img moves when press a correct choice

* Redirect to result page when player reaches the finish line (test ver.)

* Display next question

* Add TODO comments

* Configure list views and list filters on admin site

* Change timezone

* Add inline editing of associated records on admin page

* Add link to edit each question via quiz admin on admin page

* Create and config Player in Quiz

* Fix bug when config next qustion to player

* Refactor views for easily manual testing

* Update base master (merge master to this branch)

* Add dumpdata for manual testing

* Update base master (merge updated master to this branch)

* Update base master (merge updated master to this branch)

* Update base master (merge edit-question-set template from master to this branch

* Update base master (merge question-set-detail template to this branch)

* Create Player at view instead of Quiz model

* Add test models

* Add empty folder test_views

* Change class name

* Remove notation

* Add test game view

* Update base master (merge templates to this branch)

* Add real urls and views for game system

* Update base master (merge templates to this branch)

* Set Test Quiz contain 20 questions on template for manual testing

* Fix merge conflicts

* Accept change from master (try to resolve merge conflicts)

* Add urls for manual testing

* Update base master

* Apply html form on player-name.html and question-level.html

* Apply form on game.html and add button style

* Modify views to use data from request

* Change button text

* Display time duration between each question on game page

* Add view and url for updating timer

* Modify result page to display message according to player's status

* Collect and display player's correct and wrong answers

* Update base master (merge master to this branch)

* Setup player's timer for testing and update .gitignore

* Create timer(js) and display time on game page (real-time update)

* Fix timer to display smoothly when refresh game page

* Move script from game.html to script file

* Stop display time when a player reaches the finish line

* Modify models to collect time limit and player's timeout status, and modify views to handle hard level condition

* Change timer color on game.html from black to red when time is up

* Add TODO comments and delete unnecessary parts

* Clear session storage when access these pages

* Try to do overlay on game page

* Set default to time field of player model

* Update dumpdata

* Merge master to this branch

* Undo overlay on game page

* Update and edit test models

* Fix error and rename test method

* Change docstring style

* Refactor coding style and start_game view

* Add tests for views

* Add total player of quiz

* Delete useless comments

* Redirect to result.html when time is up

* Merge master to this branch

* Delete unuse migrations

* Remove unuse div

* Handle quit button

* Add field and view to handle upvote and downvote data

* Delete blank line

* Add tests for start_game and upvote_downvote views

* Add upvote and downvote links

* Display top 5 recommended quizzes when player selects quiz

* Link leaderboard button to leaderboard.html

* Edit indent

* Add home button on leaderboard.html

* Handle empty input in player_name view

* Delete TODO comment

* Set default value for html from

* Modify quit_game view (don't delete test player)

* Fix upvote and downvote test

+++ quizer_game/views.py
@@ -165,7 +165,6 @@ def update_game(request, player_id, quiz_id, selected_difficulty):
 
     # check time for hard level
     # player can still play game but player won't be ranked on leaderboard
-    # TODO prevent player from answering after time's up
     if selected_difficulty == DIFFICULTY['hard']:
         if timer.time_duration >= timer.time_limit:
             player.is_timeout = True
@@ -223,6 +222,8 @@ def update_player_position(choice, player, difficulty) -> None:
 def quit_game(request, player_id, quiz_id, selected_difficulty):
     quiz = get_object_or_404(Quiz, pk=quiz_id)
     player = quiz.player_set.get(pk=player_id)
+    if player.name in PLAYERS_FOR_TESTING:
+        return redirect(reverse('quizer_game:index'))
     player.delete()
     return redirect(reverse('quizer_game:index'))
 
@@ -253,7 +254,7 @@ def result(request, player_id, quiz_id, selected_difficulty):
     context = {'quiz': quiz, 'player': player}
     return render(request, 'quizer_game/result.html', context)
   
-
+  
 def leaderboard(request, quiz_id, selected_difficulty):
     quiz = get_object_or_404(Quiz, pk=quiz_id)
     players = quiz.player_set.filter(selected_difficulty=selected_difficulty,
@@ -358,8 +359,9 @@ def edit_data(request,quiz_id):
     messages.success(request, 'Successful saving')
     return redirect(reverse('quizer_game:edit_quiz', kwargs={'quiz_id': quiz.id}))
 
-
+  
 def quiz_index(request):
     quizzes = Quiz.objects.all()
     context = {'quizzes': quizzes}
     return render(request, 'quizer_game/quiz-index.html', context)
+
